requirements.txt
Pillow
pdf2image
prettytable
pymongo
rapidfuzz
requests
retrying
thefuzz
